feat(new-market-overview): remove legacy code
