--- src/app/[locale]/lender/components/MarketsSection/components/MarketsTables/OtherMarketsTables/interface.ts
@@ -1,34 +1,0 @@
-import { MarketAccount, TokenAmount } from "@wildcatfi/wildcat-sdk"
-
-import { BorrowerWithName } from "@/app/[locale]/borrower/hooks/useBorrowerNames"
-import { SmallFilterSelectItem } from "@/components/SmallFilterSelect"
-import { getMarketStatusChip, MarketStatus } from "@/utils/marketStatus"
-import { getMarketTypeChip } from "@/utils/marketType"
-
-export type LenderOtherMarketsTableModel = {
-  id: string
-  chainId?: number
-  status: ReturnType<typeof getMarketStatusChip>
-  term: ReturnType<typeof getMarketTypeChip>
-  name: string
-  borrower: string | undefined
-  borrowerAddress: string | undefined
-  asset: string
-  debt: TokenAmount | undefined
-  apr: number
-  withdrawalBatchDuration: number
-  isSelfOnboard: boolean
-  button?: string
-  capacityLeft: TokenAmount
-}
-
-export type LenderOtherMarketsTableProps = {
-  marketAccounts: MarketAccount[]
-  borrowers: BorrowerWithName[]
-  isLoading: boolean
-  filters: {
-    nameFilter: string
-    assetFilter: SmallFilterSelectItem[]
-    statusFilter: MarketStatus[]
-  }
-}
