--- src/app/[locale]/lender/components/MarketsSection/components/MarketsTables/interface.ts
@@ -1,17 +1,0 @@
-import { GridColDef } from "@mui/x-data-grid"
-import { Market, MarketAccount } from "@wildcatfi/wildcat-sdk"
-
-import { SmallFilterSelectItem } from "@/components/SmallFilterSelect"
-import { MarketStatus } from "@/utils/marketStatus"
-
-export type MarketsTablesProps = {
-  marketAccounts: MarketAccount[]
-  isLoading: boolean
-  filters: {
-    nameFilter: string
-    assetFilter: SmallFilterSelectItem[]
-    statusFilter: MarketStatus[]
-  }
-}
-
-export type TypeSafeColDef<T> = GridColDef & { field: keyof T }
