+++ src/app/[locale]/lender/components/TopMarketsSection/index.tsx
@@ -3,7 +3,7 @@
 import { useMemo, useState } from "react"
 import * as React from "react"
 
-import { Box, Button, Skeleton, Typography } from "@mui/material"
+import { Box, Button, Typography } from "@mui/material"
 import {
   DataGrid,
   GridRow,
@@ -21,7 +21,6 @@ import { useTranslation } from "react-i18next"
 
 import { TypeSafeColDef } from "@/app/[locale]/borrower/components/MarketsSection/сomponents/MarketsTables/interface"
 import { LinkCell } from "@/app/[locale]/borrower/components/MarketsTables/style"
-import { LenderOtherMarketsTableModel } from "@/app/[locale]/lender/components/MarketsSection/components/MarketsTables/OtherMarketsTables/interface"
 import { TopMarketSectionSelect } from "@/app/[locale]/lender/components/TopMarketSectionSelect"
 import { useLenderMarketsContext } from "@/app/[locale]/lender/context"
 import { MarketStatusChip } from "@/components/@extended/MarketStatusChip"
@@ -96,6 +95,23 @@ export const DataGridSx = {
     minHeight: "66px",
     height: "auto",
   },
+}
+
+export type LenderOtherMarketsTableModel = {
+  id: string
+  chainId?: number
+  status: ReturnType<typeof getMarketStatusChip>
+  term: ReturnType<typeof getMarketTypeChip>
+  name: string
+  borrower: string | undefined
+  borrowerAddress: string | undefined
+  asset: string
+  debt: TokenAmount | undefined
+  apr: number
+  withdrawalBatchDuration: number
+  isSelfOnboard: boolean
+  button?: string
+  capacityLeft: TokenAmount
 }
 
 const MarketLinkRow = (props: GridRowProps) => (
